Added star

+++ map-app/src/data/PointData.java
@@ -1,10 +1,10 @@
 package data;
 
 public class PointData implements IShapeData {
-    public float x;
-    public float y;
+    public double x;
+    public double y;
 
-    public PointData(float x, float y) {
+    public PointData(double x, double y) {
         this.x = x;
         this.y = y;
     }
