+++ map-app/src/shapes/Point.java
@@ -5,7 +5,7 @@ import services.draw.DrawPoint;
 
 public class Point extends Shape {
 
-    public Point(float x, float y) {
+    public Point(double x, double y) {
         setData(new PointData(x, y));
         setDrawStrategy(new DrawPoint());
     }
